project delete works now. Added new method for response rate tables

+++ app/Http/Controllers/ProjectController.php
@@ -279,7 +279,13 @@ class ProjectController extends AppBaseController
 
     private function down($project)
     {
-        Schema::drop($project->dbname);
+        if (Schema::hasTable($project->dbname)) {
+            Schema::drop($project->dbname);
+        }
+
+        if (Schema::hasTable($project->dbname . '_double')) {
+            Schema::drop($project->dbname . '_double');
+        }
     }
 
     private function createTable($dbname, $project, $fields)
